requirements.txt
python-dotenv==1.0.1
wheel==0.42.0
requests==2.32.3
pydub==0.25.1
langchain==0.3.0
flask==3.0.0
flask-cors==5.0.0
python-dotenv==1.0.1
bs4==0.0.2
azure-cognitiveservices-speech==1.40.0
langchain==0.3.0

langchain-community==0.3.0
langchain-core==0.3.1
langchain-google-genai==2.0.0
langchain-groq==0.2.0
langchain-text-splitters==0.3.0
langchainhub==0.1.21
langchain-google-genai==2.0.0
faiss-cpu
waitress
ffmpeg-python
gunicorn
wave
